+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Adds widgets:
     - `Checkbox`
-    - `Dropdown`
+    - `Dropdown`, capable of single-select or multi-select
     - `MenuBar`, accessible via `Application().set_menubar`.
     - `StatusBar`, accessible via `Application().status = ...`.
 - Adds `Widget.defer_set` as an alias of `lambda: w.tkset(**kwargs)`.
@@ -14,6 +14,7 @@
 - Adds `DynamicWidget` middlelayer to implement uniform interface for any widget that contains a tk variable.
 - `theming.Style` now takes `active_background` and `active_foreground` fields (instead of just `active` in place of `active_background`).
 - `theming.Style` now also provides `.tkdict() -> dict[str, str]` to allow for `widget.tkset(**style.tkdict())`.
+- `Widget.tkset` now continues to update `self._tk_kwargs` to allow objects (`Dropdown`) to be bound multiple times while continuing to inherit the correct settings.
 
 ## v0.1.0